finrobot_demo/requirements.txt
pyautogen>=0.2.19
ipython
huggingface_hub
matplotlib

# financial libraries
finnhub-python
yfinance
mplfinance
backtrader
sec_api

# data handling
numpy>=2.1.0
pandas>=2.1.0
pyPDF2
reportlab
pyautogen[retrievechat]

# packaging
setuptools>=41.4.0
wheel>=0.33.6

# historical data
praw
tushare
pandas_datareader

pdfkit>=1.0.0
# marker-pdf is optional and heavy; excluded for Python 3.13 minimal demo
aiohttp>=3.9.0
langchain>=0.1.20
nltk==3.8.1
ratelimit==2.2.1
requests>=2.31.0
scikit-learn>=1.6.0
starlette>=0.37.2
tenacity>=8.3.0
tqdm>=4.66.1
typing_extensions>=4.9.0
unstructured>=0.12.0
